requirements.txt
aniso8601==8.0.0
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
Click==7.0
dnspython==1.16.0
Flask==1.1.1
flask-marshmallow==0.11.0
flask-restplus==0.13.0
idna==2.10
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
MarkupSafe==1.1.1
marshmallow==3.5.0
mysql-connector-python==8.0.19
numpy==1.19.1
pandas==1.1.1
protobuf==3.6.1
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
requests==2.24.0
six==1.14.0
urllib3==1.25.10
Werkzeug==0.16.0
zipp==3.0.0
